Make instalation of AI packages optional and create a requieremnts file without AI packages
