+++ requirements_no_ai.txt
@@ -1,0 +1,146 @@
+absl-py==2.1.0
+antlr4-python3-runtime==4.9.3
+apptools==5.3.0
+astunparse==1.6.3
+black==24.8.0
+Brotli==1.1.0
+brukerapi==0.1.9
+build==1.2.2
+cachetools==5.5.0
+certifi==2024.8.30
+cfgv==3.4.0
+charset-normalizer==3.3.2
+click==8.1.7
+contourpy==1.3.0
+cycler==0.12.1
+deepdiff==8.0.1
+dipy==1.9.0
+distlib==0.3.8
+einops==0.8.0
+envisage==7.0.3
+filelock==3.16.0
+flatbuffers==24.3.25
+fonttools==4.53.1
+fsspec==2024.9.0
+gast==0.4.0
+google-auth==2.37.0
+google-auth-oauthlib==1.0.0
+google-pasta==0.2.0
+grpcio==1.66.1
+h5py==3.11.0
+huggingface-hub==0.27.1
+hydra-core==1.3.2
+identify==2.6.1
+idna==3.10
+imageio==2.35.1
+inflate64==1.0.0
+iniconfig==2.0.0
+itk==5.4.0
+itk-core==5.4.0
+itk-elastix==0.21.0
+itk-filtering==5.4.0
+itk-io==5.4.0
+itk-numerics==5.4.0
+itk-registration==5.4.0
+itk-segmentation==5.4.0
+Jinja2==3.1.4
+joblib==1.4.2
+kiwisolver==1.4.7
+lazy_loader==0.4
+libclang==18.1.1
+Markdown==3.7
+markdown-it-py==3.0.0
+MarkupSafe==2.1.5
+matplotlib==3.9.2
+mayavi==4.8.2
+mdurl==0.1.2
+ml-dtypes==0.3.2
+MouseInfo==0.1.3
+mpmath==1.3.0
+multivolumefile==0.2.3
+mypy-extensions==1.0.0
+namex==0.0.8
+networkx==3.3
+nibabel==5.2.1
+nodeenv==1.9.1
+nptyping==2.5.0
+numpy==1.24.3
+oauthlib==3.2.2
+omegaconf==2.3.0
+opencv-python==4.10.0.84
+opt-einsum==3.3.0
+optree==0.13.1
+orderly-set==5.2.2
+packaging==24.1
+pandas==2.2.2
+pathspec==0.12.1
+pillow==10.4.0
+pip-tools==7.4.1
+platformdirs==4.3.3
+pluggy==1.5.0
+pre-commit==3.8.0
+protobuf==4.25.4
+psutil==6.0.0
+py7zr==0.22.0
+pyarrow==17.0.0
+pyasn1==0.6.1
+pyasn1_modules==0.4.1
+PyAutoGUI==0.9.54
+pybcj==1.0.2
+pycryptodomex==3.20.0
+pydicom==3.0.0
+pyface==8.0.0
+PyGetWindow==0.0.9
+Pygments==2.18.0
+PyMsgBox==1.0.9
+pynrrd==1.1.1
+pyobjc-core==10.3.1
+pyobjc-framework-Cocoa==10.3.1
+pyobjc-framework-Quartz==10.3.1
+pyparsing==3.1.4
+pyperclip==1.9.0
+pyppmd==1.1.0
+pyproject_hooks==1.1.0
+PyQt6==6.7.1
+PyQt6-Qt6==6.7.2
+PyQt6_sip==13.8.0
+PyRect==0.2.0
+PyScreeze==1.0.1
+pystackreg==0.2.7
+pytest==8.3.3
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.1
+pytweening==1.2.0
+pytz==2024.2
+PyWavelets==1.7.0
+PyYAML==6.0.2
+pyzstd==0.16.1
+requests==2.32.3
+requests-oauthlib==2.0.0
+rich==13.8.1
+rsa==4.9
+rubicon-objc==0.4.9
+safetensors==0.4.5
+scikit-image==0.24.0
+scikit-learn==1.5.2
+scipy==1.14.1
+setuptools-scm==8.1.0
+six==1.16.0
+sympy==1.13.2
+termcolor==2.4.0
+texttable==1.7.0
+threadpoolctl==3.5.0
+tifffile==2024.8.30
+timm==1.0.9
+tqdm==4.66.5
+traits==6.4.3
+traitsui==8.0.0
+trx-python==0.3
+typing_extensions==4.5.0
+tzdata==2024.1
+urllib3==2.2.3
+virtualenv==20.26.4
+vtk==9.3.1
+Werkzeug==3.0.4
+wrapt==1.14.1
+xarray==2024.9.0
